+++ source/parlex/src/match.cpp
@@ -6,7 +6,7 @@ namespace parlex {
 namespace detail {
 
 match::match(uint32_t documentPosition, uint16_t recognizerIndex, uint32_t const consumedCharacterCount) : match_class(documentPosition, recognizerIndex), consumed_character_count(consumedCharacterCount) {}
-match::match(struct match_class const & matchClass, size_t const consumedCharacterCount) : match_class(matchClass), consumed_character_count(consumedCharacterCount) { }
+match::match(struct match_class const & matchClass, uint32_t const consumedCharacterCount) : match_class(matchClass), consumed_character_count(consumedCharacterCount) { }
 
 match & match::operator=(match && move) noexcept {
 	this->~match();
